move PolylineDataManager to model package
